use the new system in gitlab
