+++ livecheck/special/gitlab.py
@@ -1,10 +1,10 @@
 import requests
 import re
 from urllib.parse import urlparse, quote
-from datetime import datetime
 from loguru import logger
 
-from ..utils.portage import is_version_development
+from ..settings import LivecheckSettings
+from ..utils.portage import get_last_version
 
 __all__ = ("get_latest_gitlab_package")
 
@@ -12,21 +12,19 @@ __all__ = ("get_latest_gitlab_package")
 VERSIONS = 40
 
 
-def extract_domain_and_namespace(url: str) -> tuple[str, str]:
+def extract_domain_and_namespace(url: str) -> tuple[str, str, str]:
     parsed = urlparse(url)
-    domain = parsed.netloc
     path = parsed.path.strip('/')
     if '/-/' in path:
         path = path.split('/-/')[0]
 
-    return domain, path
+    return parsed.netloc, path, path.split('/')[-1]
 
 
-def get_latest_gitlab_package(url: str,
-                              development: bool = False,
-                              restrict_version: str = '') -> tuple[str, str, str]:
+def get_latest_gitlab_package(url: str, ebuild: str, development: bool, restrict_version: str,
+                              settings: LivecheckSettings) -> tuple[str, str]:
 
-    domain, path_with_namespace = extract_domain_and_namespace(url)
+    domain, path_with_namespace, repo = extract_domain_and_namespace(url)
     base_api_url = f"https://{domain}/api/v4"
     encoded_path = quote(path_with_namespace, safe='')
     try:
@@ -36,7 +34,7 @@ def get_latest_gitlab_package(url: str,
         tags_data = tags_response.json()
     except requests.exceptions.HTTPError as e:
         logger.error(f"URL error: {e}")
-        return '', '', ''
+        return '', ''
 
     results = []
     for tag in tags_data:
@@ -46,25 +44,13 @@ def get_latest_gitlab_package(url: str,
         if match:
             cleaned_name = match.group(1)
 
-        date_str = tag.get("commit", {}).get("created_at", "")
-        try:
-            dt = datetime.fromisoformat(date_str.replace("Z", "+00:00"))
-            formatted_date = dt.strftime("%Y%m%d")
-        except:
-            formatted_date = date_str[:10]
-
         results.append({
             "tag": cleaned_name,
             "id": tag.get("commit", {}).get("id", ""),
-            "date": formatted_date
         })
 
-    for result in results:
-        version = result["tag"]
-        if not version.startswith(restrict_version):
-            continue
-        if not is_version_development(version) or development:
-            return version, result["id"], result["date"]
+    result = get_last_version(results, repo, ebuild, development, restrict_version, settings)
+    if result:
+        return result['version'], result["id"]
 
-    logger.debug('No updated hash found')
-    return '', '', ''
+    return '', ''
